backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.3.1            # via django
django-filter==2.4.0      # via -r requirements.in
django==3.1.3             # via -r requirements.in, django-filter, djangorestframework
djangorestframework==3.12.2  # via -r requirements.in
markdown==3.3.3           # via -r requirements.in
psycopg2-binary==2.8.6    # via -r requirements.in
pytz==2020.4              # via django
sqlparse==0.4.1           # via django
